Add navigation and more KUKSA.VAL signals
- Add naviagtion widget to the app using openstreetmap
  and openrouteservice
- Add more kuksa signals
- Add more indicators for new signals
- Change the color of guages
- App layout locked aspect ratio 16:9

This add the navigation widget to the app, which shows the map and
path between two places using openstreetmap as map provider and
openrouteservice as API for path between two points.
Color of guages is less monochrome for better distinction between
safe and danger limits.
More signals and indicators are added, which are present in the QT
cluster app.
App layout is changed because app was causing errror for some
aspect ratio. Now aspect ratio is locked to 16:9.
Sometime turn indicators was blinking out of sync. To solve that
now same aniamtion controller is used for both turn signals